refactor : folderItem 분리
