+++ components/module-folder/src/main/java/depromeet/ohgzoo/iam/folder/FolderService.java
@@ -1,6 +1,5 @@
 package depromeet.ohgzoo.iam.folder;
 
-
 public interface FolderService {
     FolderResponse createFolder(Long memberId, FolderCreateRequest request);
 
@@ -8,5 +7,4 @@ public interface FolderService {
 
     FolderResponse updateFolder(Long memberId, Long folderId, FolderUpdateRequest request);
 
-    void createFolderItem(Long memberId, Long folderId, FolderItemCreateRequest request);
 }
