+++ components/module-folder/src/main/java/depromeet/ohgzoo/iam/folder/FolderServiceImpl.java
@@ -1,8 +1,5 @@
 package depromeet.ohgzoo.iam.folder;
 
-import depromeet.ohgzoo.iam.folder.exception.ExistedNameException;
-import depromeet.ohgzoo.iam.folder.exception.InvalidUserException;
-import depromeet.ohgzoo.iam.folder.exception.NotExistsFolderException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -12,7 +9,6 @@ import java.util.Optional;
 @RequiredArgsConstructor
 public class FolderServiceImpl implements FolderService {
     private final FolderRepository folderRepository;
-    private final FolderItemRepository folderItemRepository;
 
 
     public FolderResponse createFolder(Long memberId, FolderCreateRequest request) {
@@ -43,15 +39,5 @@ public class FolderServiceImpl implements FolderService {
 
         folder.updateName(request.getFolderName());
         return new FolderResponse(folder.getId(), folder.getName());
-    }
-
-    public void createFolderItem(Long memberId, Long folderId, FolderItemCreateRequest request) {
-        Folder folder = folderRepository.findById(folderId)
-                .orElseThrow(NotExistsFolderException::new);
-
-        FolderItem folderItem = new FolderItem(request.getFirstCategory(), request.getSecondCategory(), request.getContent(), request.getTags(), request.getDisclosure());
-        folderItemRepository.save(folderItem);
-        folder.addFolderItem(folderItem);
-        folder.changeCoverImg(folderItem.getFirstCategory());
     }
 }
