+++ components/module-folder/src/main/java/depromeet/ohgzoo/iam/folder/InvalidUserException.java
@@ -1,4 +1,4 @@
-package depromeet.ohgzoo.iam.folder.exception;
+package depromeet.ohgzoo.iam.folder;
 
 public class InvalidUserException extends RuntimeException {
 }
