+++ components/module-folder/src/main/java/depromeet/ohgzoo/iam/folder/ValidationException.java
@@ -1,4 +1,4 @@
-package depromeet.ohgzoo.iam.folder.exception;
+package depromeet.ohgzoo.iam.folder;
 
 public class ValidationException extends RuntimeException{
 }
