+++ components/module-folder/src/main/java/depromeet/ohgzoo/iam/folder/folderItem/FolderItemCreateRequest.java
@@ -1,5 +1,7 @@
-package depromeet.ohgzoo.iam.folder;
+package depromeet.ohgzoo.iam.folder.folderItem;
 
+import depromeet.ohgzoo.iam.folder.FirstCategory;
+import depromeet.ohgzoo.iam.folder.SecondCategory;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
