+++ components/module-folder/src/main/java/depromeet/ohgzoo/iam/folder/folderItem/FolderItemService.java
@@ -1,0 +1,6 @@
+package depromeet.ohgzoo.iam.folder.folderItem;
+
+public interface FolderItemService {
+    void createFolderItem(Long memberId, Long folderId, FolderItemCreateRequest request);
+
+}
