+++ components/module-folder/src/test/java/depromeet/ohgzoo/iam/folder/SpyFolderService.java
@@ -1,15 +1,16 @@
 package depromeet.ohgzoo.iam.folder;
 
+import depromeet.ohgzoo.iam.folder.folderItem.FolderItemCreateRequest;
+
 public class SpyFolderService implements FolderService {
     public Long argument_memberId;
     public FolderResponse createFolder_returnValue;
     public Long deleteFolder_argumentId;
     public FolderUpdateRequest updateFolder_argumentRequest;
     public FolderCreateRequest createFolder_argumentRequest;
-    public FolderItemCreateRequest createFolderItem_argumentRequest;
     public Long updateFolder_argumentMemberId;
     public Long updateFolder_argumentFolderId;
-    public Long createFolderItem_argumentFolderId;
+
     public FolderResponse updateFolder_returnValue;
 
 
@@ -34,10 +35,4 @@ public class SpyFolderService implements FolderService {
         return updateFolder_returnValue;
     }
 
-    @Override
-    public void createFolderItem(Long memberId, Long folderId, FolderItemCreateRequest request){
-        argument_memberId = memberId;
-        createFolderItem_argumentRequest=request;
-        createFolderItem_argumentFolderId=memberId;
-    }
 }
